% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_document_pair_distances.R
\name{calculate_document_pair_distances}
\alias{calculate_document_pair_distances}
\title{Document Distances}
\usage{
calculate_document_pair_distances(document_term_matrix, document_indicies,
  distance_method = "cosine")
}
\arguments{
\item{document_term_matrix}{A sparse matrix object of class
"simple_triplet_matrix", or a dense matrix object, with documents as rows and
vocabulary entries as columns.}

\item{document_indicies}{A numeric vector of length two (document_a_row_index,
document_b_row_index), or a list object with each entry containing a vector
of length two as described above.}

\item{distance_method}{Can be one of "cosine" or "euclidean". Defaults to
"cosine".}
}
\value{
A vector of document pair distances
}
\description{
Calculate distances between pairs of documents.
}

