% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_directory_name.R
\name{check_directory_name}
\alias{check_directory_name}
\title{A function to ensure that a directory name is in the proper format to be pasted together with a file name. It adds a trailling / if necessary.}
\usage{
check_directory_name(directory)
}
\arguments{
\item{directory}{A string containing a directory path that we wish to check.}
}
\value{
A string object that is a directory name with a trailing slash, which is what will be required if we wish to paste it and a file name together.
}
\description{
A function to ensure that a directory name is in the proper format to be pasted together with a file name. It adds a trailling / if necessary.
}
