% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean_document_text.R
\name{clean_document_text}
\alias{clean_document_text}
\title{A function which cleans the raw text of a document provided either as a single string, a vector of strings, or a column of a data.frame.}
\usage{
clean_document_text(text, remove_regex = "[^a-zA-Z\\\\s]")
}
\arguments{
\item{text}{The raw text of a document the user wishes to clean. Can be supplied as either a single string, a vector of strings, or a column from a data.frame.}

\item{remove_regex}{A regular expression specifying the characters the user would like to exclude from the final text string. This function works by replacing those terms with spaces and then splitting the resulting string on those spaces. Defaults to removing all characters that are not uper or lowercase letters or spaces (as a regex, this is "[^a-zA-Z\\s]").}
}
\value{
A document-term vector with ordering preserved.
}
\description{
A function which cleans the raw text of a document provided either as a single string, a vector of strings, or a column of a data.frame.
}

