% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_tf_idf_scalings.R
\name{compare_tf_idf_scalings}
\alias{compare_tf_idf_scalings}
\title{A function that performs a bunch of different forms of TF-IDF scaling to a
document-term matrix.}
\usage{
compare_tf_idf_scalings(document_term_matrix)
}
\arguments{
\item{document_term_matrix}{A document-term matrix as a dense or simple
triplet matrix, with documents as rows, and vocabulary terms as columns. Note
that internally, this function will turn the matrix into a dense matrix, so
this function should only be used with moderately sized matrices.}
}
\value{
A list object containing a number of document-term matrices, each
with a different form of TF-IDF scaling applied. Each list entry contains a
list object with the rescaled doc-term matrix as the second entry, and a
description of the method applied as the first entry.
}
\description{
A function that performs a bunch of different forms of TF-IDF scaling to a
document-term matrix.
}

