% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documentation.R
\docType{data}
\name{congress_bills}
\alias{congress_bills}
\title{All versions of the first 20 bills introduced in the House and Senate in the 103rd Congress.}
\format{A character vector of length 79.}
\usage{
data(congress_bills)
}
\description{
A character vector containing 79 entries, one per document, for all versions
for the first 20 bills introduced in the U.S. House of Representatives, and
all versions of the first 21 bills introduced in the U.S. Senate in the 103rd
session of Congress (1993-1994).
}
\keyword{datasets}
