% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_quanteda_to_slam.R
\name{convert_quanteda_to_slam}
\alias{convert_quanteda_to_slam}
\title{A function to convert a quanteda dfm object to a slam::simple_triplet_matrix.}
\usage{
convert_quanteda_to_slam(quanteda_dfm)
}
\arguments{
\item{quanteda_dfm}{A dfm object generated by the quanteda::dfm() function.}
}
\description{
A function to convert a quanteda dfm object to a slam::simple_triplet_matrix.
}
