% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corenlp.R
\name{corenlp}
\alias{corenlp}
\title{Runs Stanford CoreNLP on a collection of documents}
\usage{
corenlp(documents = NULL, document_directory = NULL,
  delete_intermediate_files = TRUE, version = "3.5.2")
}
\arguments{
\item{documents}{An optional list of character vectors or a vector of strings, with one entry per dcument. These documents will be run through CoreNLP.}

\item{document_directory}{An optional directory path to a directory contianing only .txt files (one per document) to be run through CoreNLP. Cannot be supplied in addition to the 'documents' argument.}

\item{delete_intermediate_files}{Logical indicating whether intermediate files produced by CoreNLP should be deleted. Defaults to TRUE, but can be set to FALSE and the xml output of CoreNLP will be saved.}

\item{version}{The version of Core-NLP to download. Defaults to '3.5.2'.}
}
\value{
Does not return anything.
}
\description{
Runs Stanford CoreNLP on a collection of documents
}

