% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corenlp_blocked.R
\name{corenlp_blocked}
\alias{corenlp_blocked}
\title{Runs Stanford CoreNLP on a collection of .txt files and processes them in blocks of a specified size, saving intermediate results to disk. Designed to function on very large corpora.}
\usage{
corenlp_blocked(output_directory, document_directory, file_list = NULL,
  block_size = 1000, syntactic_parsing = FALSE,
  coreference_resolution = FALSE, additional_options = "",
  return_raw_output = FALSE, version = "3.5.2", parallel = FALSE,
  cores = 1, first_block = NULL, last_block = NULL)
}
\arguments{
\item{output_directory}{The path to a directory where the user would like CoreNLP output to be stored. Output will be saved to this directory in .Rdata files named CoreNLP_Output_1.Rdata ... CoreNLP_Output_N.Rdata}

\item{document_directory}{A directory path to a directory contianing .txt files (one per document) to be run through CoreNLP.}

\item{file_list}{An optional list of .txt files to be used. Can be useful if the user only wants to process a subset of documents in the directory such as when the corpus is extremely large.}

\item{block_size}{The number of docuemnts to be processed at a time. Defaults to 1000.}

\item{syntactic_parsing}{Logical indicating whether syntactic parsing should be included as an option. Defaults to FALSE. Caution, enabling this argument may greatly increase runtime. If TRUE, output will automatically be return in raw format.}

\item{coreference_resolution}{Logical indicating whether coreference resolution should be included as an option. Defaults to FALSE. Caution, enabling this argument may greatly increase runtime. If TRUE, output will automatically be return in raw format.}

\item{additional_options}{An optional string specifying additional options for CoreNLP. May cause unexpected behavior, use at your own risk!}

\item{return_raw_output}{Defaults to FALSE, if TRUE, then CoreNLP output is not parsed and raw list objects are returned.}

\item{version}{The version of Core-NLP to download. Defaults to '3.5.2'. Newer versions of CoreNLP will be made available at a later date.}

\item{parallel}{Logical indicating whether CoreNLP should be run in parallel.}

\item{cores}{The number of cores to be used if CoreNLP is being run in parallel.}

\item{first_block}{Used to run CoreNLP on certain block ranges.}

\item{last_block}{Used to run CoreNLP on certain block ranges.}
}
\value{
Does not return anything, saves all output to disk.
}
\description{
Runs Stanford CoreNLP on a collection of .txt files and processes them in blocks of a specified size, saving intermediate results to disk. Designed to function on very large corpora.
}

