% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice_coefficient_diff_table.R
\name{dice_coefficient_diff_table}
\alias{dice_coefficient_diff_table}
\title{Lines In Both Documents via Dice Coefficients}
\usage{
dice_coefficient_diff_table(document_1, document_2, dice_matching_results,
  output_string = FALSE, compare_consecutive_line_pairs = TRUE)
}
\arguments{
\item{document_1}{A vector of strings (one per line or one per sentence), or
a list of vectors of tokens  (one per line or one per sentence).}

\item{document_2}{Same as document_1, will be used for comparison.}

\item{dice_matching_results}{Output from dice_coefficient_line_matching(),
with dice_matrix included.}

\item{output_string}{Logical indicating whether a string vector should be
returned. If TRUE, then all output is in string vector, if FALSE, the output
is concatenated to the screen.}

\item{compare_consecutive_line_pairs}{Logical indicating whether consecutive
pairs of lines should be compared. Defaults to TRUE.}
}
\value{
A character vector (optionally).
}
\description{
Calculate termwise Deice coefficeints for pairs of lines and
determine which lines are in both documents
}

