% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distinct_words.R
\name{distinct_words}
\alias{distinct_words}
\title{A function to find (semi)-distinct words in a list of term vectors.}
\usage{
distinct_words(word_vector_list, threshold = 1)
}
\arguments{
\item{word_vector_list}{A list of character vectors we wish to find distinctive words in.}

\item{threshold}{An integer > 0 indicating the number of times a word must appear more than to be included in the vector we return. Defaults to threshold = 1, meaning all words that appear 1 or less times in the other term vectors we pass in will be removed from them before they are compared against the current vector. In this way we can get pseudo-distinct words, perhaps preventing us from removing really distinctive words that appear only threshold or less times in most term vectors, but lots of times in one vector in particular.}
}
\value{
A list of distinct word vectors.
}
\description{
A function to find (semi)-distinct words in a list of term vectors.
}

