% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Data_Documentation.R
\docType{data}
\name{document_term_count_list}
\alias{document_term_count_list}
\title{Document Term Count List: Conressional Bills}
\format{A list object with 5 fields.}
\usage{
data(document_term_count_list)
}
\description{
A list of five document term count vectors drawn from a sample of bills introduced
in congress in the 1980's. These represent the ounts of unique terms in the
document_term_vector_list object.
}
\keyword{datasets}

