% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documentation.R
\docType{data}
\name{document_term_vector_list}
\alias{document_term_vector_list}
\title{Document Term Vector List: Conressional Bills}
\format{A list object with 5 fields.}
\usage{
data(document_term_vector_list)
}
\description{
A list of five document term vectors drawn from a sample of bills introduced
in congress in the 1980's. These only represent unique terms but are meant to
serve as an illustration.
}
\keyword{datasets}

