% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_corenlp.R
\name{download_corenlp}
\alias{download_corenlp}
\title{Checks the java version on your computer and downloads Stanford CoreNLP jar files for use with this package.}
\usage{
download_corenlp(version = "3.5.2")
}
\arguments{
\item{version}{The version of Core-NLP to download. Defaults to '3.5.2'}
}
\value{
Does not return anything.
}
\description{
Checks the java version on your computer and downloads Stanford CoreNLP jar files for use with this package.
}
