% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_mallet.R
\name{download_mallet}
\alias{download_mallet}
\title{Checks the java version on your computer and downloads MALLET jar files for use with this package.}
\usage{
download_mallet()
}
\value{
Does not return anything.
}
\description{
Checks the java version on your computer and downloads MALLET jar files for use with this package.
}

