% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_metrics.R
\name{edit_metrics}
\alias{edit_metrics}
\title{Calculate Edit Metrics Between Two Document Versions}
\usage{
edit_metrics(document_1, document_2, ngram_sizes = c(1:50))
}
\arguments{
\item{document_1}{A vector of strings (one per line or one per sentence), or
a list of vectors of tokens  (one per line or one per sentence).}

\item{document_2}{Same as document_1, will be used for comparison.}

\item{ngram_sizes}{A numeric vector of N-Gram lengths for us in calculating
Dice coefficients.}
}
\value{
A data.frame with Dice coefficients based on different N-Gram lengths.
}
\description{
Calculate Scope and Granularity of Document Edits.
}

