% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{feature_selection}
\alias{feature_selection}
\title{A function that implements a number of feature selection methods for finding top
words which distinguish between two classes.}
\usage{
feature_selection(contingency_table, rows_to_compare = c(1, 2), alpha = 1,
  method = "informed_Dirichlet")
}
\arguments{
\item{contingency_table}{A contingency table generated by the `contingency_table()` function.}

\item{rows_to_compare}{A numeric vector containing the indicies of two rows
in the contingency table we wish to compare against eachother. Defaults to
c(1,2)}

\item{alpha}{The Dirichlet hyperparameter to be used if method =
"informed_Dirichlet". Suggested value is the average number of terms that
appear in a document. If a small value is selected, then more (globally)
common terms may be selected as top words. Increasing the value will select
for less globally common words. Defaults to 1 (not usually a good choice for
most analyses).}

\item{method}{Currently only "informed_Dirichlet" is availalbe.
"informed_Dirichlet" implements the model described in section 3.5.1 of
Monroe, B. L., Colaresi, M. P., and Quinn, K. M. (2008). Fightin' words:
Lexical feature selection and evaluation for identifying the content of
political conflict. Political Analysis, 16, 372–403.}
}
\value{
A list object containing two dataframes (one for each comparison
category) with ranked top words. All words included in each dataset obtain
a z-score greater in magnitude than 1.96.
}
\description{
A function that implements a number of feature selection methods for finding top
words which distinguish between two classes.
}

