% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_threshold.R
\name{frequency_threshold}
\alias{frequency_threshold}
\title{A function to frequency threshold a vector of strings.}
\usage{
frequency_threshold(word_vector, threshold = 1)
}
\arguments{
\item{word_vector}{A character vector of words we wish to frequency threshold.}

\item{threshold}{An integer > 0 indicating the number of times a word must appear more than to be included in the vector we return. Defaults to threshold = 1, meaning all words that appear 1 or less times in the vector we pass in will be removed.}
}
\value{
A word vector with all words appearing 'threshold' or less times removed.
}
\description{
A function to frequency threshold a vector of strings.
}

