% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_contingency_table.R
\name{generate_contingency_table}
\alias{generate_contingency_table}
\title{Generates a contingency table from user-specified document covariates and a document term matrix.}
\usage{
generate_contingency_table(metadata, document_term_matrix, vocabulary = NULL,
  variables_to_use = NULL, threshold = 0)
}
\arguments{
\item{metadata}{A data.frame containing document covariates.}

\item{document_term_matrix}{A documents x vocabulary matrix with counts of unique words in each document. Can be a dense or sparse matrix.}

\item{vocabulary}{A character vector corresponding to the columns of the document word matrix. If NULL, the column names of doc_word_matrix will be used. Defaults to NULL.}

\item{variables_to_use}{Defaults to NULL in which case all columns of the metadata data frame will be used. Otherwise can be specified as a vector of column indexes or column names.}

\item{threshold}{Defaults to 0, the number of times a unique value of a variable must appear in order to be included in the returned list object. Allows the user to ignore very infrequent values.}
}
\value{
A contingency table.
}
\description{
Generates a contingency table from user-specified document covariates and a document term matrix.
}

