% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_document_term_matrix.R
\name{generate_document_term_matrix}
\alias{generate_document_term_matrix}
\title{A function to generate a document term matrix from a list of document term vectors.}
\usage{
generate_document_term_matrix(document_term_vector_list, vocabulary = NULL,
  document_term_count_list = NULL, return_sparse_matrix = FALSE)
}
\arguments{
\item{document_term_vector_list}{A list of term vectors, one per document, that we wish to turn into a document term matrix.}

\item{vocabulary}{An optional vocabulary vector which will be used to form the document term matrix. Defaults to NULL, in which case a vocabulary vector will be generated internally.}

\item{document_term_count_list}{A list of vectors of word counts can optionally be provided, in which case we will aggregate over them. This can be useful if we wish to store documents in a memory efficent way. Defaults to NULL.}

\item{return_sparse_matrix}{Defualts to FALSE, in whih case a normal dense matrix is returned. If TRUE, then a sparse matrix object generated by the slam library is returned. A sparse matrix representation is also used in the C++ code if this is set to TRUE, which can result in drastic memory savings.}
}
\value{
A dense document term matrix object with the vocabulary as column names.
}
\description{
A function to generate a document term matrix from a list of document term vectors.
}
