% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_document_term_vector_list.R
\name{generate_document_term_vector_list}
\alias{generate_document_term_vector_list}
\title{A function to generate a document term vector list in a standard format and save it to file.}
\usage{
generate_document_term_vector_list(input, return_result = FALSE)
}
\arguments{
\item{input}{A list of term vectors, one per document, that we wish to save.}

\item{return_result}{A logical value indicating whether the resulting list object should be returned.}
}
\value{
A document term vector list.
}
\description{
A function to generate a document term vector list in a standard format and save it to file.
}

