% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_document_term_vectors.R
\name{generate_document_term_vectors}
\alias{generate_document_term_vectors}
\title{A function to generate document term vectors from a variety of inputs.}
\usage{
generate_document_term_vectors(input, data_type = c("string", "term vector",
  "raw text", "csv"), data_directory = NULL,
  tokenization_method = c("RegEx", "CoreNLP"), return_result = FALSE,
  csv_separator = ",", csv_word_column = NULL, csv_count_column = NULL)
}
\arguments{
\item{input}{A list of strings, term vectors, raw documents, or csv files you wish to turn into document term vectors.}

\item{data_type}{The type of data provided to the function.}

\item{data_directory}{Optional argument specifying where the data is stored.}

\item{tokenization_method}{Will eventually implement both an R-based and a Stanford CoreNLP based tokenization strategy. Currently not available.}

\item{return_result}{A logical value indicating whether the resulting list object should be returned.}

\item{csv_separator}{Defaults to "," but can be set to "\t" for tab separated values.}

\item{csv_word_column}{If you are providing one csv file per document, then you must specify the index of the column that contains the words. Defaults to NULL.}

\item{csv_count_column}{For memory efficiency, you may want to store only the counts of unique words in csv files. If your data include counts, then you must specify the index of the column that contains the counts. Defaults to NULL.}
}
\value{
A document term vector list.
}
\description{
A function to generate document term vectors from a variety of inputs.
}

