% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unique_values_and_counts.R
\name{get_unique_values_and_counts}
\alias{get_unique_values_and_counts}
\title{Find unique values and the counts of those variables for a set of variables in a data.frame. Useful in PMI analysis and for exploring document metadata.}
\usage{
get_unique_values_and_counts(metadata, variable_names = NA, threshold = 0)
}
\arguments{
\item{metadata}{A data.frame containing document covariates.}

\item{variable_names}{A string or vector of strings givign the column names of covariates we would like to get unique values for}

\item{threshold}{Defaults to 0, the number of times a}
}
\value{
A vector (or list of vectors) of unique variable values that occur more than "threshold" times.
}
\description{
Find unique values and the counts of those variables for a set of variables in a data.frame. Useful in PMI analysis and for exploring document metadata.
}

