% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kill_zombies.R
\name{kill_zombies}
\alias{kill_zombies}
\title{A function which takes no arguments and kills zombie R processes if the user is using a UNIX based machine}
\usage{
kill_zombies()
}
\description{
A function which takes no arguments and kills zombie R processes if the user is using a UNIX based machine
}

