% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multi_plot.R
\name{multi_plot}
\alias{multi_plot}
\title{An implementation of matplot with nice coloring and autmatic legend generation}
\usage{
multi_plot(data, plot_columns = NULL, legend_location = "bottomleft",
  xlabel = "", ylabel = "", category = "", output_pdf = FALSE,
  pdf_name = NULL, height = 5, width = 8)
}
\arguments{
\item{data}{A data.frame or matrix we wish to plot.}

\item{plot_columns}{A numeric vector of column indexes we wish to plot. Defaults to NULL in which case all columns are plotted.}

\item{legend_location}{Defaults to "bottomleft" but can be changed to any of "left", "topleft", "top", "topright", "right", "bottomright", "bottom".}

\item{xlabel}{The label we wish to give the X axis. Defaults to no label.}

\item{ylabel}{The label we wish to give the Y axis. Defaults to no label.}

\item{category}{The label we wish to give to the legend box.}

\item{output_pdf}{Defaults to FALSE. If TRUE, then the plot is output to a pdf with pdf_name to the current working directory.}

\item{pdf_name}{The name we wish to give out pdf. No .pdf extension is required as it is automatically appended.}

\item{height}{The height of the optional .pdf file generated by this function. Defaults to 5 (inches).}

\item{width}{The width of the optional .pdf file generated by this function. Defaults to 8 (inches).}
}
\value{
A dense document term matrix object with the vocabulary as column names.
}
\description{
An implementation of matplot with nice coloring and autmatic legend generation
}

