% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutual_information.R
\name{mutual_information}
\alias{mutual_information}
\title{Mutual Information}
\usage{
mutual_information(joint_dist, normalized = FALSE,
  non_zero_column_entries = NULL)
}
\arguments{
\item{joint_dist}{A matrix of class "simple_triplet_matrix"
or a standard dense matrix.}

\item{normalized}{Defaults to FALSE. If TRUE, then the function expects a
matrix which sums to one. WARNING -- this function will not check to see if
the matrix sums  to one if TRUE (for speed reasons).}

\item{non_zero_column_entries}{Defaults to NULL, otherwise, a vector of
length equal to the number of rows in the joint_dist recording the number of
non-zero column entries for that row. Can greatly speed up computation. Must
be supplied if normalized = TRUE.}
}
\value{
The mutual information of the joint distribution.
}
\description{
Calculate Mutual Information
}
