% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngram_sequence_matching.R
\name{ngram_sequence_matching}
\alias{ngram_sequence_matching}
\title{N-Gram Sequence Matching}
\usage{
ngram_sequence_matching(document_1, document_2, ngram_size,
  use_hashmap = FALSE, tokenized_strings_provided = FALSE)
}
\arguments{
\item{document_1}{A string (or a character vector) representing the earlier
document version.}

\item{document_2}{A string (or a character vector) representing the later
document version.}

\item{ngram_size}{The length of n-grams to be compared}

\item{use_hashmap}{Defaults to FALSE. If TRUE, then a hashmap is used for
faster lookup and comparisons.}

\item{tokenized_strings_provided}{Defaults to FALSE. If TRUE, then
pre-tokenized strings are expected as character vectors.}
}
\value{
A List object.
}
\description{
Calculates the positions of n-grams in two document versions
which match an ngram in the other version.
}
