% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngram_sequence_plot.R
\name{ngram_sequnce_plot}
\alias{ngram_sequnce_plot}
\title{N-Gram Sequence Matching}
\usage{
ngram_sequnce_plot(ngram_sequnce_results, document = 1, columns = NULL,
  include_title_legend = TRUE, return_plot_object = FALSE,
  custom_title = NULL, match_color = "#2B83BAC8",
  non_match_color = "#FDAE61C8")
}
\arguments{
\item{ngram_sequnce_results}{A List object produced by the
`ngram_sequence_matching()` function.}

\item{document}{Can be 1 or 2, the index of the document to plot.}

\item{columns}{Defaults to NA, in which case the number of columns will be
set automatically. Can be set to any integer number by the user.}

\item{include_title_legend}{Defaults to TRUE in which cas a title and legend
are include in the plot. Can be set to FALSE for multi-plotting.}

\item{return_plot_object}{Logical defaults to FALSE. If TRUE, then the plot
is not printed, only returned.}

\item{custom_title}{Optional custom plot title.}

\item{match_color}{Defaults to "#2B83BAC8" (light blue).}

\item{non_match_color}{Defaults to "#FDAE61C8" (orange).}
}
\value{
A plot
}
\description{
Calculates the positions of n-grams in two document versions
which match an ngram in the other version.
}
