% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_by_counts.R
\name{order_by_counts}
\alias{order_by_counts}
\title{A function to generate an ordered word count dataframe from a raw vector of words.}
\usage{
order_by_counts(word_vector, weighting = NULL)
}
\arguments{
\item{word_vector}{A character vector of words we wish to count and order.}

\item{weighting}{a vector of weightings to be given to counts of a particular string. Defaults to NULL meaning no weightings will be used.}
}
\value{
A data.frame with 'word' and a 'count' fields, with rows ordered by word frequency.
}
\description{
A function to generate an ordered word count dataframe from a raw vector of words.
}
