% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documentation.R
\docType{data}
\name{Processed_Text}
\alias{Processed_Text}
\title{Twenty bills tokenized and tagged by CORENLP}
\format{A list object containing 20 data.frames.}
\usage{
data(Processed_Text)
}
\description{
A list of twenty congressional bills from 1991 that have been preprocessed
and tokenized by the corenlp() function. For use in testing the ngrams()
function.
}
\keyword{datasets}

