% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sparse_to_dense_matrix.R
\name{sparse_to_dense_matrix}
\alias{sparse_to_dense_matrix}
\title{A function to convert a slam::simple_triplet_matrix sparse matrix object to a dense matrix object.}
\usage{
sparse_to_dense_matrix(sparse_matrix)
}
\arguments{
\item{sparse_matrix}{A sparse matrix object generated by the slam library.}
}
\value{
A dense matrix
}
\description{
A function to convert a slam::simple_triplet_matrix sparse matrix object to a dense matrix object.
}

