% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unlist_and_concatenate.R
\name{unlist_and_concatenate}
\alias{unlist_and_concatenate}
\title{A function to unlist and concatenate a subset of a matrix/data.frame}
\usage{
unlist_and_concatenate(dataframe_chunk)
}
\arguments{
\item{dataframe_chunk}{A subset of a dataframe or matrix we wish to stick into a vector}
}
\value{
A vector of words
}
\description{
A function to unlist and concatenate a subset of a matrix/data.frame
}

