% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corenlp.R
\name{corenlp}
\alias{corenlp}
\title{Runs Stanford CoreNLP on a collection of documents}
\usage{
corenlp(documents = NULL, document_directory = NULL, version = "3.5.2")
}
\arguments{
\item{documents}{An optional list of character vectors or a vector of strings, with one entry per dcument. These documents will be run through CoreNLP.}

\item{document_directory}{An optional directory path to a directory contianing only .txt files (one per document) to be run through CoreNLP. Cannot be supplied in addition to the 'documents' argument.}

\item{version}{The version of Core-NLP to download. Defaults to '3.5.2'.}
}
\value{
Does not return anything.
}
\description{
Runs Stanford CoreNLP on a collection of documents
}

