% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corenlp.R
\name{corenlp}
\alias{corenlp}
\title{Runs Stanford CoreNLP on a collection of documents}
\usage{
corenlp(documents = NULL, document_directory = NULL,
  delete_intermediate_files = TRUE, syntactic_parsing = FALSE,
  coreference_resolution = FALSE, additional_options = "",
  return_raw_output = FALSE, version = "3.5.2")
}
\arguments{
\item{documents}{An optional list of character vectors or a vector of strings, with one entry per dcument. These documents will be run through CoreNLP.}

\item{document_directory}{An optional directory path to a directory contianing only .txt files (one per document) to be run through CoreNLP. Cannot be supplied in addition to the 'documents' argument.}

\item{delete_intermediate_files}{Logical indicating whether intermediate files produced by CoreNLP should be deleted. Defaults to TRUE, but can be set to FALSE and the xml output of CoreNLP will be saved.}

\item{syntactic_parsing}{Logical indicating whether syntactic parsing should be included as an option. Defaults to FALSE. Caution, enabling this argument may greatly increase runtime. If TRUE, output will automatically be return in raw format.}

\item{coreference_resolution}{Logical indicating whether coreference resolution should be included as an option. Defaults to FALSE. Caution, enabling this argument may greatly increase runtime. If TRUE, output will automatically be return in raw format.}

\item{additional_options}{An optional string specifying additional options for CoreNLP. May cause unexpected behavior, use at your own risk!}

\item{return_raw_output}{Defaults to FALSE, if TRUE, then CoreNLP output is not parsed and raw list objects are returned.}

\item{version}{The version of Core-NLP to download. Defaults to '3.5.2'. Newer versions of CoreNLP will be made available at a later date.}
}
\value{
Returns a list of data.frame objects, one per document, where each row is a token observation (in order)
}
\description{
Runs Stanford CoreNLP on a collection of documents
}
\examples{
\dontrun{
directory <- system.file("extdata", package = "SpeedReader")[1]
Tokenized <- corenlp(
     documents = NULL,
     document_directory = directory,
     delete_intermediate_files = TRUE,
     syntactic_parsing = FALSE,
     coreference_resolution =FALSE,
     additional_options = "",
     return_raw_output = FALSE)
}
}

