% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice_coefficient_line_matching.R
\name{dice_coefficient_line_matching}
\alias{dice_coefficient_line_matching}
\title{Lines In Both Documents via Dice Coefficients}
\usage{
dice_coefficient_line_matching(document_1, document_2, threshold = 0.8,
  return_dice_matrix = TRUE, compare_consecutive_line_pairs = TRUE,
  whole_document = FALSE)
}
\arguments{
\item{document_1}{A vector of strings (one per line or one per sentence), or
a list of vectors of tokens  (one per line or one per sentence).}

\item{document_2}{Same as document_1, will be used for comparison.}

\item{threshold}{A value between 0 and 1 that denotes the Dice coefficient
threshold for considdering a line as being in both documents.}

\item{return_dice_matrix}{Logical indicating whether the Dice coefficient
matrix should be returned. Defualts to TRUE.}

\item{compare_consecutive_line_pairs}{Logical indicating whether consecutive
pairs of lines should be compared. Defaults to TRUE.}

\item{whole_document}{Logical, defaults to FALSE. If TRUE, then all lines are
combined and full documents are compared.}
}
\value{
A list with two vectors, each giving the inidices of the lines
in document 1/2 that are in the other document based on the dice coefficient
threshold.
}
\description{
Calculate termwise Deice coefficeints for pairs of lines and
determine which lines are in both documents
}

