% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{feature_selection}
\alias{feature_selection}
\title{A function that implements a number of feature selection methods for finding top
words which distinguish between two classes.}
\usage{
feature_selection(contingency_table, rows_to_compare = c(1, 2), alpha = 1,
  method = c("informed_Dirichlet", "TF-IDF"), maximum_top_words = 5000)
}
\arguments{
\item{contingency_table}{A contingency table generated by the `contingency_table()` function.}

\item{rows_to_compare}{A numeric vector containing the indicies of two rows
in the contingency table we wish to compare against eachother. Defaults to
c(1,2). Currenlty only works for two category comparisons.}

\item{alpha}{The Dirichlet hyperparameter to be used if method =
"informed_Dirichlet". Suggested value is the average number of terms that
appear in a document. If a small value is selected, then more (globally)
common terms may be selected as top words. Increasing the value will select
for less globally common words. Defaults to 1 (not usually a good choice for
most analyses).}

\item{method}{Defaults to "informed_Dirichlet", which implements the model
described in section 3.5.1 of Monroe et al. "Fightin Words...". Can also be
"TF-IDF", in which case TF-IDF ranking is used (following equation 2 in
"Fightin Words").}

\item{maximum_top_words}{Controls the maximum number of top words returned in
each category. Defaults to 5000.}
}
\value{
A list object containing two dataframes (one for each comparison
category) with ranked top words. All words included in each dataset obtain
a z-score greater in magnitude than 1.96.
}
\description{
A function that implements a number of feature selection methods for finding top
words which distinguish between two classes.
}

