% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_document_term_vectors.R
\name{generate_document_term_vectors}
\alias{generate_document_term_vectors}
\title{A function to generate document term vectors from a variety of inputs.}
\usage{
generate_document_term_vectors(input, data_type = c("string", "term vector",
  "raw text", "csv"), data_directory = NULL,
  tokenization_method = c("RegEx", "CoreNLP"), output_type = c("return",
  "save", "return and save"), output_name = NULL, csv_separator = ",",
  csv_word_column = NULL, csv_count_column = NULL, csv_header = FALSE,
  keep_sequence = FALSE)
}
\arguments{
\item{input}{A list of strings, term vectors, raw documents, or csv files you wish to turn into document term vectors.}

\item{data_type}{The type of data provided to the function.}

\item{data_directory}{Optional argument specifying where the data is stored.}

\item{tokenization_method}{Will eventually implement both an R-based and a Stanford CoreNLP based tokenization strategy. Currently not available.}

\item{output_type}{A string value indicating whether the resulting list object should be returned ("return"), saved to a file in the current working directory with name output_name.Rdata ("save"), of both ("save and return"). Defaults to "return".}

\item{output_name}{The file name we wish to give the output document term vector list object generated by this function if specifying a value of "save" or "return and save" for output_type. You do not need to provide the .Rdata extension as this is appended automatically. Defaults to NULL.}

\item{csv_separator}{Defaults to "," but can be set to "*backslash*t" for tab separated values.}

\item{csv_word_column}{If you are providing one csv file per document, then you must specify the index of the column that contains the words. Defaults to NULL.}

\item{csv_count_column}{For memory efficiency, you may want to store only the counts of unique words in csv files. If your data include counts, then you must specify the index of the column that contains the counts. Defaults to NULL.}

\item{csv_header}{Logical indicating whether the csv files provided have a header. Defaults to FALSE.}

\item{keep_sequence}{Logical indicating whether document term vectors should be condensed and counts (FALSE) or whether the full sequence should be maintained for storage (TRUE). Defaults to FALSE as this can be a much more memory efficient representation.}
}
\value{
A document term vector list.
}
\description{
A function to generate document term vectors from a variety of inputs.
}

