% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_distribution_distance.R
\name{reference_distribution_distance}
\alias{reference_distribution_distance}
\title{Reference distribtuion distances}
\usage{
reference_distribution_distance(category_reference_distribution,
  document_term_matrix, inverse_frequency_weighting = TRUE)
}
\arguments{
\item{category_reference_distribution}{A simple_triplet_matrix where each row
represents the distribution over terms in a particular category. These can
be normalized or raw counts.}

\item{document_term_matrix}{A simple_triplet_matrix where each row represents
a document and each column, a term in the vocabulary. The columns in both
matrices should match up.}

\item{inverse_frequency_weighting}{If TRUE, then distances are weighted by
the inverse of the term's aggregate count in the document term
matrix. This means that differences in more frequently occuring terms will
have less weight than those for less frequently appearing terms. Defaults to
TRUE.}
}
\value{
A dataframe with distances of each document to each reference
distribution. The last column indicates the closest reference distribtuion
for each document.
}
\description{
Calculates the euclidean distance (up to a proportionality)
between document term distributions and a set of reference distributions.
}
