% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_doc_term_parallel.R
\name{sparse_doc_term_parallel}
\alias{sparse_doc_term_parallel}
\title{Only to be used internally. A function to generate a sparse large document term matrix in parallel.}
\usage{
sparse_doc_term_parallel(file, vocabulary)
}
\arguments{
\item{file}{The path to a block of document term vectors.}

\item{vocabulary}{This is set internally inside the generate_sparse_large_document_term_matrix() function.}
}
\value{
A sparse document term matrix object.
}
\description{
Only to be used internally. A function to generate a sparse large document term matrix in parallel.
}
